# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Hermes Local AI Runtime API
  version: 0.2.0-dev
  description: >
    G-05 loopback contract. Native `/api/v1` is normative; `/v1` is a
    compatibility adapter. Operations marked `planned` are not implemented.
  license:
    name: Apache-2.0
servers:
  - url: http://127.0.0.1:8850
    description: G-05 loopback example
security:
  - bearerAuth: []
paths:
...
  /api/v1/jobs:
    post:
      x-implementation-status: implemented-g05
      operationId: submitJob
      parameters:
        - in: header
          name: Idempotency-Key
          schema: {type: string, minLength: 8, maxLength: 128}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../schemas/job-request.schema.json"
      responses:
        "200":
          description: Idempotent replay of the original job submission
        "202":
          description: Asynchronous job accepted
          headers:
            Location:
              required: true
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAccepted"
        "400": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "413": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
  /api/v1/jobs/{job_id}:
    get:
      x-implementation-status: implemented-g05
      operationId: getJob
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Job state or result
        "404": {$ref: "#/components/responses/Error"}
...
  /api/v1/jobs/{job_id}/result:
    get:
      x-implementation-status: implemented-g05
      operationId: getJobResult
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Immutable result with timing and provenance
          content:
            application/json:
              schema:
                $ref: "../schemas/job-result.schema.json"
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
  /api/v1/uploads:
    post:
      x-implementation-status: implemented-g05
      operationId: uploadBoundedInput
      description: >
        Streams a bounded raw body to a bounded volatile memory store. Content
        is never written to disk or logged and disappears at process shutdown.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: {type: string, format: binary}
      responses:
        "201": {description: Upload metadata; never the payload}
        "413": {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
  parameters:
    JobId:
      in: path
      name: job_id
      required: true
      schema: {type: string, pattern: "^job_[A-Za-z0-9_-]+$"}
  schemas:
...
    JobAccepted:
      type: object
      additionalProperties: false
      required: [job_id, capability, capability_version, profile, status, route_id, status_url, result_url]
      properties:
        job_id: {type: string}
        capability: {type: string}
        capability_version: {type: string}
        profile: {type: string, enum: [fast, balanced, accurate]}
        status: {const: queued}
        route_id: {type: string}
        status_url: {type: string}
        result_url: {type: string}
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, retryable, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            retryable: {type: boolean}
            request_id: {type: string, pattern: "^req_[A-Za-z0-9_-]+$"}
            retry_after_seconds: {type: integer, minimum: 0}
            details: {type: array, items: {type: object, additionalProperties: true}}
  responses:
    Error:
      description: Stable runtime error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"